+++ enershelf/map/layers.py
@@ -1,18 +1,16 @@
 import json
-import os
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from itertools import product
 from typing import List, Optional
 
-from django.db.models import IntegerField, BooleanField
+from django.contrib.gis.db.models import Model
+from django.db.models import IntegerField, BooleanField, ObjectDoesNotExist
+from raster.models import RasterLayer as RasterModel
 
 from config.settings.base import USE_DISTILLED_MVTS
-from .config import MAX_ZOOM, MIN_ZOOM, REGIONS, ZOOM_LEVELS, MAX_DISTILLED_ZOOM
+from .config import LAYER_STYLES, MAX_ZOOM, MIN_ZOOM, REGIONS, ZOOM_LEVELS, MAX_DISTILLED_ZOOM
 
 from . import models
-
-with open(os.path.join(os.path.dirname(__file__), "../static/styles/layer_styles.json"), mode="rb",) as f:
-    LAYER_STYLES = json.loads(f.read())
 
 
 def get_color(source_layer):
@@ -23,33 +21,48 @@ def get_opacity(source_layer):
     return LAYER_STYLES[source_layer]["paint"]["fill-opacity"]
 
 
-ELECTRICITY = [
-    # {
-    #     "source": "grid",
-    #     "color": get_color("grid"),
-    #     "model": models.Grid,
-    #     "name": "Grids",
-    #     "name_singular": "Grid",
-    #     "description": "Electricity grids",
-    # },
-    {
-        "source": "nightlight",
-        "color": get_color("nightlight"),
-        "model": models.Nightlight,
-        "name": "Nightlights",
-        "name_singular": "Nightlight",
-        "description": "See nightlights",
-    },
+@dataclass
+class VectorLayerData:
+    source: str
+    color: str
+    model: Model.__class__
+    name: str
+    name_singular: str
+    description: str
+    popup_fields: list = field(default_factory=list)
+
+
+@dataclass
+class RasterLayerData:
+    source: str
+    filepath: str
+    legend: str
+    model: RasterModel.__class__
+    name: str
+    name_singular: str
+    description: str
+
+
+ELECTRICITY: list = [
+    VectorLayerData(
+        source="nightlight",
+        color=get_color("nightlight"),
+        model=models.Nightlight,
+        name="Nightlights",
+        name_singular="Nightlight",
+        description="See nightlights",
+    )
 ]
-HOSPITALS = [
-    {
-        "source": "built_up_areas",
-        "color": get_color("built_up_areas"),
-        "model": models.BuiltUpAreas,
-        "name": "Built Up Areas",
-        "name_singular": "Built Up Area",
-        "description": "See cluster",
-        "popup_fields": [
+
+HOSPITALS: list = [
+    VectorLayerData(
+        source="built_up_areas",
+        color=get_color("built_up_areas"),
+        model=models.BuiltUpAreas,
+        name="Built Up Areas",
+        name_singular="Built Up Area",
+        description="See cluster",
+        popup_fields=[
             "id",
             "area",
             "population",
@@ -58,50 +71,42 @@ HOSPITALS = [
             "distance_to_light",
             "district_name",
         ],
-    },
-    {
-        "source": "settlements",
-        "color": get_color("settlements"),
-        "model": models.Settlements,
-        "name": "Settlements",
-        "name_singular": "Settlement",
-        "description": "See cluster",
-        "popup_fields": ["id", "area", "population", "number_of_hospitals", "distance_to_grid", "distance_to_light"],
-    },
-    {
-        "source": "hamlets",
-        "color": get_color("hamlets"),
-        "model": models.Hamlets,
-        "name": "Hamlets",
-        "name_singular": "Hamlet",
-        "description": "See cluster",
-        "popup_fields": ["id", "area", "population", "number_of_hospitals"],
-    },
-    {
-        "source": "hospital",
-        "color": "red",
-        "model": models.Hospitals,
-        "name": "Hospitals",
-        "name_singular": "Hospital",
-        "description": "See nightlights test",
-        "popup_fields": [
-            "id",
-            "name",
-            "type",
-            "town",
-            "ownership",
-            "population_per_hospital",
-            "catchment_area_hospital",
-        ],
-    },
-    # {
-    #     "source": "hospital_simulated",
-    #     "color": "red",
-    #     "model": models.HospitalsSimulated,
-    #     "name": "Simulated Hospitals",
-    #     "name_singular": "Simulated hospital",
-    #     "description": "See nightlights test",
-    #     "popup_fields": [
+    ),
+    VectorLayerData(
+        source="settlements",
+        color=get_color("settlements"),
+        model=models.Settlements,
+        name="Settlements",
+        name_singular="Settlement",
+        description="See cluster",
+        popup_fields=["id", "area", "population", "number_of_hospitals", "distance_to_grid", "distance_to_light"],
+    ),
+    VectorLayerData(
+        source="hamlets",
+        color=get_color("hamlets"),
+        model=models.Hamlets,
+        name="Hamlets",
+        name_singular="Hamlet",
+        description="See cluster",
+        popup_fields=["id", "area", "population", "number_of_hospitals"],
+    ),
+    VectorLayerData(
+        source="hospital",
+        color="red",
+        model=models.Hospitals,
+        name="Hospitals",
+        name_singular="Hospital",
+        description="See nightlights test",
+        popup_fields=["id", "name", "type", "town", "ownership", "population_per_hospital", "catchment_area_hospital"],
+    ),
+    # VectorLayerData(
+    #     source="hospital_simulated",
+    #     color="red",
+    #     model=models.HospitalsSimulated,
+    #     name="Simulated Hospitals",
+    #     name_singular="Simulated hospital",
+    #     description="See nightlights test",
+    #     popup_fields=[
     #         "id",
     #         "name",
     #         "type",
@@ -110,10 +115,21 @@ HOSPITALS = [
     #         "population_per_hospital",
     #         "catchment_area_hospital",
     #     ],
-    # },
+    # ),
 ]
-LAYERS_DEFINITION = ELECTRICITY + HOSPITALS
-LAYERS_CATEGORIES = {"Electricty": ELECTRICITY, "Hospitals": HOSPITALS}
+SOLAR: list = [
+    RasterLayerData(
+        source="solar",
+        filepath="Gha_Yearly_Solar_horizontal_irradiation.tif",
+        legend="solar",
+        model=RasterModel,
+        name="Solar",
+        name_singular="Solar",
+        description="See nightlights test",
+    ),
+]
+LAYERS_DEFINITION = ELECTRICITY + HOSPITALS + SOLAR
+LAYERS_CATEGORIES = {"Electricty": ELECTRICITY, "Hospitals": HOSPITALS, "Solar": SOLAR}
 
 
 @dataclass
@@ -139,6 +155,13 @@ class Layer:
 
 
 @dataclass
+class RasterLayer:
+    id: str
+    source: str
+    type: str
+
+
+@dataclass
 class Popup:
     source: str
     layer_id: str
@@ -161,18 +184,37 @@ def get_layer_setups(layer):
 def get_dynamic_sources():
     sources = []
     for layer in LAYERS_DEFINITION:
-        if not hasattr(layer["model"], "setup"):
+        if not hasattr(layer.model, "setup"):
             continue
         for combination in get_layer_setups(layer):
             mvt_str = "-".join(combination)
             filter_str = "&".join(map(lambda x: f"setup__{x}", combination))
             sources.append(
                 Source(
-                    name=f"{layer['source']}-{mvt_str}",
+                    name=f"{layer.source}-{mvt_str}",
                     type="vector",
-                    tiles=[f"{layer['source']}_mvt/{{z}}/{{x}}/{{y}}/?{filter_str}"],
+                    tiles=[f"{layer.source}_mvt/{{z}}/{{x}}/{{y}}/?{filter_str}"],
                 )
             )
+    return sources
+
+
+def get_raster_sources(distilled=False):
+    sources = []
+    for layer in LAYERS_DEFINITION:
+        if not issubclass(layer.model, RasterModel):
+            continue
+        try:
+            raster_id = RasterModel.objects.get(name=layer.source).id
+        except ObjectDoesNotExist:
+            continue
+        sources.append(
+            Source(
+                name=f"{layer.source}",
+                type="raster",
+                tiles=[f"raster/tiles/{raster_id}/{{z}}/{{x}}/{{y}}.png?legend={layer.legend}"],
+            )
+        )
     return sources
 
 
@@ -193,6 +235,7 @@ if USE_DISTILLED_MVTS:
             Source(name="static", type="vector", tiles=["static_mvt/{z}/{x}/{y}/"]),
             Source(name="static_distilled", type="vector", tiles=["static/mvts/{z}/{x}/{y}/static.mvt"],),
         ]
+        + get_raster_sources()
         + get_dynamic_sources()
     )
 else:
@@ -200,6 +243,7 @@ else:
     ALL_SOURCES = (
         [Source(name=region, type="vector", tiles=[f"{region}_mvt/{{z}}/{{x}}/{{y}}/"]) for region in REGIONS]
         + [Source(name="static", type="vector", tiles=["static_mvt/{z}/{x}/{y}/"])]
+        + get_raster_sources()
         + get_dynamic_sources()
     )
 
@@ -242,33 +286,45 @@ REGION_LAYERS = (
     ]
 )
 
+RASTER_LAYERS = [
+    RasterLayer(id=layer.source, source=layer.source, type="raster",)
+    for layer in LAYERS_DEFINITION
+    if issubclass(layer.model, RasterModel)
+]
+
 POPUPS = []
 STATIC_LAYERS = []
 for layer in LAYERS_DEFINITION:
-    if hasattr(layer["model"], "setup"):
+    if issubclass(layer.model, RasterModel):
+        continue
+    if hasattr(layer.model, "setup"):
         continue
     for suffix in SUFFIXES:
-        layer_id = f"fill-{layer['source']}{suffix}"
+        layer_id = f"fill-{layer.source}{suffix}"
         STATIC_LAYERS.append(
             Layer(
                 id=layer_id,
-                color=layer["color"],
-                description=layer["description"],
+                color=layer.color,
+                description=layer.description,
                 minzoom=MAX_DISTILLED_ZOOM + 1 if suffix == "" and USE_DISTILLED_MVTS else MIN_ZOOM,
                 maxzoom=MAX_ZOOM if suffix == "" else MAX_DISTILLED_ZOOM + 1,
-                name=layer["name"],
-                style=layer["source"],
+                name=layer.name,
+                style=layer.source,
                 source=f"static{suffix}",
-                source_layer=layer["source"],
+                source_layer=layer.source,
                 type="static",
             )
         )
-        if "popup_fields" in layer:
+        if layer.popup_fields:
             popup_fields = {}
-            for field in layer["popup_fields"]:
-                label = getattr(layer["model"], field).field.verbose_name if hasattr(layer["model"], field) else field
-                popup_fields[label] = field
-            POPUPS.append(Popup(layer["source"], layer_id, json.dumps(popup_fields)))
+            for popup_field in layer.popup_fields:
+                label = (
+                    getattr(layer.model, popup_field).field.verbose_name
+                    if hasattr(layer.model, popup_field)
+                    else popup_field
+                )
+                popup_fields[label] = popup_field
+            POPUPS.append(Popup(layer.source, layer_id, json.dumps(popup_fields)))
 
 # Sort popups according to prio:
 POPUP_PRIO = ["hospital", "hospital_simulated"]  # from high to low prio
@@ -276,19 +332,19 @@ POPUPS = sorted(POPUPS, key=lambda x: len(POPUP_PRIO) if x.source not in POPUP_P
 
 DYNAMIC_LAYERS = [
     Layer(
-        id=f"fill-{layer['source']}-{'-'.join(combination)}",
-        color=layer["color"],
-        description=layer["description"],
+        id=f"fill-{layer.source}-{'-'.join(combination)}",
+        color=layer.color,
+        description=layer.description,
         minzoom=MIN_ZOOM,
         maxzoom=MAX_ZOOM,
-        name=layer["name"],
-        style=layer["source"],
-        source=f"{layer['source']}-{'-'.join(combination)}",
-        source_layer=layer["source"],
+        name=layer.name,
+        style=layer.source,
+        source=f"{layer.source}-{'-'.join(combination)}",
+        source_layer=layer.source,
         type="static",
     )
     for layer in LAYERS_DEFINITION
-    if hasattr(layer["model"], "setup")
+    if hasattr(layer.model, "setup")
     for combination in get_layer_setups(layer)
 ]