+++ enershelf/utils/data_processing.py
@@ -1,6 +1,9 @@
+import json
 import os
 
-from config.settings.base import ROOT_DIR
+from raster.models import Legend, RasterLayer as RasterModel
+
+from config.settings.base import DATA_DIR
 from enershelf.utils.ogr_layer_mapping import RelatedModelLayerMapping
 from enershelf.map.models import (
     Region,
@@ -13,6 +16,8 @@ from enershelf.map.models import (
     Hamlets,
     Nightlight,
 )
+from enershelf.map.layers import LAYERS_DEFINITION
+from enershelf.map.config import LAYER_STYLES
 
 REGIONS = [
     Country,
@@ -37,9 +42,9 @@ def load_regions(regions=None, verbose=True):
             continue
         print(f"Upload data for region '{region.__name__}'")
         if hasattr(region, "data_folder"):
-            data_path = os.path.join(ROOT_DIR, "enershelf", "data", region.data_folder, f"{region.data_file}.gpkg")
+            data_path = os.path.join(DATA_DIR, region.data_folder, f"{region.data_file}.gpkg")
         else:
-            data_path = os.path.join(ROOT_DIR, "enershelf", "data", f"{region.data_file}.gpkg")
+            data_path = os.path.join(DATA_DIR, f"{region.data_file}.gpkg")
         region_model = Region(layer_type=region.__name__.lower())
         region_model.save()
         instance = RelatedModelLayerMapping(
@@ -57,16 +62,45 @@ def load_data(models=None, verbose=True):
             continue
         print(f"Upload data for model '{model.__name__}'")
         if hasattr(model, "data_folder"):
-            data_path = os.path.join(ROOT_DIR, "enershelf", "data", model.data_folder, f"{model.data_file}.gpkg")
+            data_path = os.path.join(DATA_DIR, model.data_folder, f"{model.data_file}.gpkg")
         else:
-            data_path = os.path.join(ROOT_DIR, "enershelf", "data", f"{model.data_file}.gpkg")
+            data_path = os.path.join(DATA_DIR, f"{model.data_file}.gpkg")
         instance = RelatedModelLayerMapping(
             model=model, data=data_path, mapping=model.mapping, layer=model.layer, transform=4326,
         )
         instance.save(strict=True, verbose=verbose)
 
 
+def load_raster(layers=LAYERS_DEFINITION):
+    for layer in layers:
+        if not issubclass(layer.model, RasterModel):
+            continue
+        if RasterModel.objects.filter(name=layer.source).exists():
+            print(f"Skipping data for raster '{layer.name}' - Please empty raster first if you want to update data.")
+            continue
+        print(f"Upload data for raster '{layer.name}'")
+        rm = RasterModel(name=layer.source, rasterfile=layer.filepath)
+        rm.save()
+        if Legend.objects.filter(title=layer.legend).exists():
+            print(
+                f"Skipping legend '{layer.legend}' for raster '{layer.name}' - "
+                f"Please remove raster legend first if you want to update it."
+            )
+            continue
+        legend = Legend(title=layer.legend, json=json.dumps(LAYER_STYLES[layer.legend]))
+        legend.save()
+
+
 def empty_data(models=None):
     models = models or MODELS
     for model in models:
         model.objects.all().delete()
+
+
+def empty_raster(layers=None):
+    layers = layers or LAYERS_DEFINITION
+    for layer in layers:
+        if not issubclass(layer.model, RasterModel):
+            continue
+        RasterModel.objects.filter(name=layer.source).delete()
+        Legend.objects.filter(title=layer.legend).delete()